+++ source/server/GameServer/keeps/Gameobjects/Guards/Hastener.cs
@@ -84,12 +84,14 @@ namespace DOL.GS.Keeps
 			if (!base.Interact(player))
 				return false;
 
-			if (player.Realm != Realm)
+			if (!GameServer.ServerRules.IsSameRealm(this, player, true))
+			{
+				GameHastener.SendSpeedBlockMessage(player, "GameHastener.SpeedBlockedRealm");
 				return false;
+			}
 
 			TurnTo(player, 5000);
-			this.CastSpellOnOwnerAndPets(player, SkillBase.GetSpellByID(GameHastener.SPEEDOFTHEREALMID), SkillBase.GetSpellLine(GlobalSpellsLines.Realm_Spells), false);
-			player.Out.SendSpellEffectAnimation(this, player, SkillBase.GetSpellByID(935).ClientEffect, 0, false, 1);
+			GameHastener.CastSpeedOfTheRealm(this, player);
 			return true;
 		}
 		#endregion Examine/Interact Message